pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ssh-mcp"
version = "0.1.0"
description = "MCP server for executing commands on remote servers via SSH"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
dependencies = ["mcp[cli]", "paramiko", "pyyaml"]

[project.optional-dependencies]
dev = ["pytest", "black", "isort", "mypy"]

[project.scripts]
ssh-mcp = "ssh_mcp.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["ssh_mcp"]
